extremely early WIP on new cortical indexing
