+++ feagi_connector_core/src/neuron_voxel_coding/xyzp/encoders/percentage_1d_exponential.rs
@@ -77,3 +77,4 @@ impl PercentageExponentialNeuronVoxelXYZPEncoder {
 
 
 
+
