+++ feagi_connector_core/src/neuron_voxel_coding/xyzp/encoders/percentage_1d_linear.rs
@@ -79,3 +79,4 @@ impl PercentageLinearNeuronVoxelXYZPEncoder {
 
 
 
+
