+++ feagi_connector_core/src/neuron_voxel_coding/xyzp/encoders/signed_percentage_1d_exponential.rs
@@ -85,3 +85,4 @@ impl SignedPercentage1DExponentialNeuronVoxelXYZPEncoder {
 
 
 
+
