+++ feagi_connector_core/src/neuron_voxel_coding/xyzp/encoders/signed_percentage_2d_exponential.rs
@@ -94,3 +94,4 @@ impl SignedPercentage2DExponentialNeuronVoxelXYZPEncoder {
 
 
 
+
