+++ feagi_data_structures/src/templates/mod.rs
@@ -5,3 +5,4 @@
 
 pub mod sensor_types;
 pub mod motor_types;
+mod sensor_cortical_units;
