+++ feagi_data_structures/src/templates/sensor_types.rs
@@ -20,53 +20,22 @@ macro_rules! sensor_definition {
 
                 //region Percentage
 
-                #[doc = "Infrared distance sensor for object detection. Absolute (instant) change. Neurons encoded linearly"]
-                InfraredAbsoluteLinear => {
-                    friendly_name: "Infrared Sensor (Absolute Change, Linear Encoding)",
-                    snake_case_identifier: "infrared_absolute_linear",
-                    base_ascii: b"iinf00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..2),
-                    default_coder_type: Percentage_Absolute_Linear,
+                #[doc = "Infrared distance sensor for object detection."]
+                Infrared => {
+                    friendly_name: "Infrared Sensor",
+                    snake_case_identifier: "infrared",
                     wrapped_data_type: WrappedIOType::Percentage,
                     data_type: Percentage,
-                },
-
-                #[doc = "Infrared distance sensor for object detection. Absolute (instant) change. Neurons encoded fractionally exponentially"]
-                InfraredAbsoluteFractional => {
-                    friendly_name: "Infrared Sensor (Absolute Change, Linear Encoding)",
-                    snake_case_identifier: "infrared_absolute_fractional",
-                    base_ascii: b"iINF00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..2),
-                    default_coder_type: Percentage_Absolute_Fractional,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
-                #[doc = "Infrared distance sensor for object detection. Incremental change. Neurons encoded linearly"]
-                InfraredIncrementalLinear => {
-                    friendly_name: "Infrared Sensor (Absolute Change, Linear Encoding)",
-                    snake_case_identifier: "infrared_incremental_linear",
-                    base_ascii: b"Iinf00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..2),
-                    default_coder_type: Percentage_Incremental_Linear,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
-                #[doc = "Infrared distance sensor for object detection. Incremental change. Neurons encoded fractionally exponentially"]
-                InfraredIncrementalFractional => {
-                    friendly_name: "Infrared Sensor (Absolute Change, Linear Encoding)",
-                    snake_case_identifier: "infrared_incremental_fractional",
-                    base_ascii: b"IINF00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..2),
-                    default_coder_type: Percentage_Incremental_Fractional,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
+                    subtype_mappings: [
+                        (b"iinf00", (Absolute, Linear), Percentage_Absolute_Linear),
+                        (b"iINF00", (Absolute, Fractional), Percentage_Absolute_Fractional),
+                        (b"Iinf00", (Incremental, Linear), Percentage_Incremental_Linear),
+                        (b"IINF00", (Incremental, Fractional), Percentage_Incremental_Fractional),
+                    ]
                 },
 
 
-
-                #[doc = "Inverted infrared sensor that provides reverse object detection readings. Absolute (instant) change. Neurons encoded linearly"]
+                #[doc = "Inverted infrared sensor that provides reverse object detection readings."]
                 ReverseInfraredAbsoluteLinear => {
                     friendly_name: "Infrared (Inverted) Sensor (Absolute Change, Linear Encoding)",
                     snake_case_identifier: "infrared_inverted_absolute_linear",
@@ -75,133 +44,47 @@ macro_rules! sensor_definition {
                     default_coder_type: Percentage_Absolute_Linear,
                     wrapped_data_type: WrappedIOType::Percentage,
                     data_type: Percentage,
+                    subtype_mappings: [
+                        (b"iiif00", (Absolute, Linear), Percentage_Absolute_Linear),
+                        (b"iIIF00", (Absolute, Fractional), Percentage_Absolute_Fractional),
+                        (b"Iiif00", (Incremental, Linear), Percentage_Incremental_Linear),
+                        (b"IIIF00", (Incremental, Fractional), Percentage_Incremental_Fractional),
+                    ]
                 },
-
-                #[doc = "Inverted infrared sensor that provides reverse object detection readings. Absolute (instant) change. Neurons encoded fractionally exponentially"]
-                ReverseInfraredAbsoluteFractional => {
-                    friendly_name: "Infrared (Inverted) Sensor (Absolute Change, Fractional Encoding)",
-                    snake_case_identifier: "infrared_inverted_absolute_fractional",
-                    base_ascii: b"iIIF00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..2),
-                    default_coder_type: Percentage_Absolute_Fractional,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
-                #[doc = "Inverted infrared sensor that provides reverse object detection readings. Incremental change. Neurons encoded linearly"]
-                ReverseInfraredIncrementalLinear => {
-                    friendly_name: "Infrared (Inverted) Sensor (Incremental Change, Linear Encoding)",
-                    snake_case_identifier: "infrared_inverted_incremental_linear",
-                    base_ascii: b"Iiif00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..2),
-                    default_coder_type: Percentage_Incremental_Linear,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
-                #[doc = "Inverted infrared sensor that provides reverse object detection readings. Incremental change. Neurons encoded fractionally exponentially"]
-                ReverseInfraredIncrementalFractional => {
-                    friendly_name: "Infrared (Inverted) Sensor (Incremental Change, Fractional Encoding)",
-                    snake_case_identifier: "infrared_inverted_incremental_fractional",
-                    base_ascii: b"IIIF00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..2),
-                    default_coder_type: Percentage_Incremental_Fractional,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
 
 
                 #[doc = "Digital GPIO input pin for reading binary signals (high/low states). Absolute (instant) change. Neurons encoded linearly"]
                 DigitalGPIOInputAbsoluteLinear => {
                     friendly_name: "GPIO Digital Input (Absolute Change, Linear Encoding)",
                     snake_case_identifier: "gpio_digital_absolute_linear",
-                    base_ascii: b"idgp00",
                     channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..2),
                     default_coder_type: Percentage_Absolute_Linear,
                     wrapped_data_type: WrappedIOType::Percentage,
                     data_type: Percentage,
+                    subtype_mappings: [
+                        (b"idgp00", (Absolute, Linear), Percentage_Absolute_Linear),
+                        (b"iDGP00", (Absolute, Fractional), Percentage_Absolute_Fractional),
+                        (b"Idgp00", (Incremental, Linear), Percentage_Incremental_Linear),
+                        (b"IDGP00", (Incremental, Fractional), Percentage_Incremental_Fractional),
+                    ]
                 },
-
-                #[doc = "Digital GPIO input pin for reading binary signals (high/low states). Absolute (instant) change. Neurons encoded fractionally exponentially"]
-                DigitalGPIOInputAbsoluteFractional => {
-                    friendly_name: "GPIO Digital Input (Absolute Change, Fractional Encoding)",
-                    snake_case_identifier: "gpio_digital_absolute_fractional",
-                    base_ascii: b"iDGP00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..2),
-                    default_coder_type: Percentage_Absolute_Fractional,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
-                #[doc = "Digital GPIO input pin for reading binary signals (high/low states). Incremental change. Neurons encoded linearly"]
-                DigitalGPIOInputIncrementalLinear => {
-                    friendly_name: "GPIO Digital Input (Incremental Change, Linear Encoding)",
-                    snake_case_identifier: "gpio_digital_incremental_linear",
-                    base_ascii: b"Idgp00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..2),
-                    default_coder_type: Percentage_Incremental_Linear,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
-                #[doc = "Digital GPIO input pin for reading binary signals (high/low states). Incremental change. Neurons encoded fractionally exponentially"]
-                DigitalGPIOInputIncrementalFractional => {
-                    friendly_name: "GPIO Digital Input (Incremental Change, Fractional Encoding)",
-                    snake_case_identifier: "gpio_digital_incremental_fractional",
-                    base_ascii: b"IDGP00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..2),
-                    default_coder_type: Percentage_Incremental_Fractional,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
 
 
                 #[doc = "Proximity sensor for detecting nearby objects and measuring distances. Absolute (instant) change. Neurons encoded linearly"]
                 ProximityAbsoluteLinear => {
                     friendly_name: "Proximity (Absolute Change, Linear Encoding)",
                     snake_case_identifier: "proximity_absolute_linear",
-                    base_ascii: b"ipro00",
                     channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..u32::MAX),
                     default_coder_type: Percentage_Absolute_Linear,
                     wrapped_data_type: WrappedIOType::Percentage,
                     data_type: Percentage,
+                    subtype_mappings: [
+                        (b"ipro00", (Absolute, Linear), Percentage_Absolute_Linear),
+                        (b"iPRO00", (Absolute, Fractional), Percentage_Absolute_Fractional),
+                        (b"Ipro00", (Incremental, Linear), Percentage_Incremental_Linear),
+                        (b"IPRO00", (Incremental, Fractional), Percentage_Incremental_Fractional),
+                    ]
                 },
-
-                #[doc = "Proximity sensor for detecting nearby objects and measuring distances. Absolute (instant) change. Neurons encoded fractionally exponentially"]
-                ProximityAbsoluteFractional => {
-                    friendly_name: "Proximity (Absolute Change, Fractional Encoding)",
-                    snake_case_identifier: "proximity_absolute_fractional",
-                    base_ascii: b"iPRO00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..33),
-                    default_coder_type: Percentage_Absolute_Fractional,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
-                #[doc = "Proximity sensor for detecting nearby objects and measuring distances. Incremental change. Neurons encoded linearly"]
-                ProximityIncrementalLinear => {
-                    friendly_name: "Proximity (Incremental Change, Linear Encoding)",
-                    snake_case_identifier: "proximity_incremental_linear",
-                    base_ascii: b"Ipro00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..u32::MAX),
-                    default_coder_type: Percentage_Incremental_Linear,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
-                #[doc = "Proximity sensor for detecting nearby objects and measuring distances. Incremental change. Neurons encoded fractionally exponentially"]
-                ProximityIncrementalFractional => {
-                    friendly_name: "Proximity (Incremental Change, Fractional Encoding)",
-                    snake_case_identifier: "proximity_incremental_fractional",
-                    base_ascii: b"IPRO00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..33),
-                    default_coder_type: Percentage_Incremental_Fractional,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
 
 
                 #[doc = "Shock sensor for sensing 'pain'. Absolute (instant) change. Neurons encoded linearly"]
@@ -213,41 +96,13 @@ macro_rules! sensor_definition {
                     default_coder_type: Percentage_Absolute_Linear,
                     wrapped_data_type: WrappedIOType::Percentage,
                     data_type: Percentage,
+                    subtype_mappings: [
+                        (b"ishk00", (Absolute, Linear), Percentage_Absolute_Linear),
+                        (b"iSHK00", (Absolute, Fractional), Percentage_Absolute_Fractional),
+                        (b"Ishk00", (Incremental, Linear), Percentage_Incremental_Linear),
+                        (b"ISHK00", (Incremental, Fractional), Percentage_Incremental_Fractional),
+                    ]
                 },
-
-                #[doc = "Shock sensor for sensing 'pain'. Absolute (instant) change. Neurons encoded fractionally exponentially"]
-                ShockAbsoluteFractional => {
-                    friendly_name: "Shock (Absolute Change, Fractional Encoding)",
-                    snake_case_identifier: "shock_absolute_fractional",
-                    base_ascii: b"iSHK00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..2),
-                    default_coder_type: Percentage_Absolute_Fractional,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
-                #[doc = "Shock sensor for sensing 'pain'. Incremental change. Neurons encoded linearly"]
-                ShockIncrementalLinear => {
-                    friendly_name: "Shock (Incremental Change, Linear Encoding)",
-                    snake_case_identifier: "shock_incremental_linear",
-                    base_ascii: b"Ishk00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..2),
-                    default_coder_type: Percentage_Incremental_Linear,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
-                #[doc = "Shock sensor for sensing 'pain'. Incremental change. Neurons encoded fractionally exponentially"]
-                ShockIncrementalFractional => {
-                    friendly_name: "Shock (Incremental Change, Fractional Encoding)",
-                    snake_case_identifier: "shock_incremental_fractional",
-                    base_ascii: b"ISHK00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..2),
-                    default_coder_type: Percentage_Incremental_Fractional,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
 
 
                 #[doc = "Battery level sensor for monitoring power remaining. Absolute (instant) change. Neurons encoded linearly"]
@@ -259,41 +114,13 @@ macro_rules! sensor_definition {
                     default_coder_type: Percentage_Absolute_Linear,
                     wrapped_data_type: WrappedIOType::Percentage,
                     data_type: Percentage,
+                    subtype_mappings: [
+                        (b"ibat00", (Absolute, Linear), Percentage_Absolute_Linear),
+                        (b"iBAT00", (Absolute, Fractional), Percentage_Absolute_Fractional),
+                        (b"Ibat00", (Incremental, Linear), Percentage_Incremental_Linear),
+                        (b"IBAT00", (Incremental, Fractional), Percentage_Incremental_Fractional),
+                    ]
                 },
-
-                #[doc = "Battery level sensor for monitoring power remaining. Absolute (instant) change. Neurons encoded fractionally exponentially"]
-                BatteryAbsoluteFractional => {
-                    friendly_name: "Battery Gauge (Absolute Change, Fractional Encoding)",
-                    snake_case_identifier: "battery_gauge_absolute_fractional",
-                    base_ascii: b"iBAT00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..33),
-                    default_coder_type: Percentage_Absolute_Fractional,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
-                #[doc = "Battery level sensor for monitoring power remaining. Incremental change. Neurons encoded linearly"]
-                BatteryIncrementalLinear => {
-                    friendly_name: "Battery Gauge (Incremental Change, Linear Encoding)",
-                    snake_case_identifier: "battery_gauge_incremental_linear",
-                    base_ascii: b"Ibat00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..u32::MAX),
-                    default_coder_type: Percentage_Incremental_Linear,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
-                #[doc = "Battery level sensor for monitoring power remaining. Incremental change. Neurons encoded fractionally exponentially"]
-                BatteryIncrementalFractional => {
-                    friendly_name: "Battery Gauge (Incremental Change, Fractional Encoding)",
-                    snake_case_identifier: "battery_gauge_incremental_fractional",
-                    base_ascii: b"IBAT00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..33),
-                    default_coder_type: Percentage_Incremental_Fractional,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
 
 
                 #[doc = "Analog GPIO input pin for reading binary signals (high/low states). Absolute (instant) change. Neurons encoded linearly"]
@@ -305,39 +132,12 @@ macro_rules! sensor_definition {
                     default_coder_type: Percentage_Absolute_Linear,
                     wrapped_data_type: WrappedIOType::Percentage,
                     data_type: Percentage,
-                },
-
-                #[doc = "Analog GPIO input pin for reading binary signals (high/low states). Absolute (instant) change. Neurons encoded fractionally exponentially"]
-                AnalogGPIOInputAbsoluteFractional => {
-                    friendly_name: "GPIO Analog Input (Absolute Change, Fractional Encoding)",
-                    snake_case_identifier: "gpio_analog_absolute_fractional",
-                    base_ascii: b"iAGP00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..33),
-                    default_coder_type: Percentage_Absolute_Fractional,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
-                #[doc = "Analog GPIO input pin for reading binary signals (high/low states). Incremental change. Neurons encoded linearly"]
-                AnalogGPIOInputIncrementalLinear => {
-                    friendly_name: "GPIO Analog Input (Incremental Change, Linear Encoding)",
-                    snake_case_identifier: "gpio_analog_incremental_linear",
-                    base_ascii: b"Iagp00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..u32::MAX),
-                    default_coder_type: Percentage_Incremental_Linear,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
-                },
-
-                #[doc = "Analog GPIO input pin for reading binary signals (high/low states). Incremental change. Neurons encoded fractionally exponentially"]
-                AnalogGPIOInputIncrementalFractional => {
-                    friendly_name: "GPIO Analog Input (Incremental Change, Fractional Encoding)",
-                    snake_case_identifier: "gpio_analog_incremental_fractional",
-                    base_ascii: b"IAGP00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..33),
-                    default_coder_type: Percentage_Incremental_Fractional,
-                    wrapped_data_type: WrappedIOType::Percentage,
-                    data_type: Percentage,
+                    subtype_mappings: [
+                        (b"iagp00", (Absolute, Linear), Percentage_Absolute_Linear),
+                        (b"iAGP00", (Absolute, Fractional), Percentage_Absolute_Fractional),
+                        (b"Iagp00", (Incremental, Linear), Percentage_Incremental_Linear),
+                        (b"IAGP00", (Incremental, Fractional), Percentage_Incremental_Fractional),
+                    ]
                 },
 
                 //endregion
@@ -353,42 +153,13 @@ macro_rules! sensor_definition {
                     default_coder_type: SignedPercentage_Absolute_Linear,
                     wrapped_data_type: WrappedIOType::SignedPercentage,
                     data_type: SignedPercentage,
+                    subtype_mappings: [
+                        (b"isvp00", (Absolute, Linear), SignedPercentage_Absolute_Linear),
+                        (b"iSVP00", (Absolute, Fractional), SignedPercentage_Absolute_Fractional),
+                        (b"Isvp00", (Incremental, Linear), SignedPercentage_Incremental_Linear),
+                        (b"ISVP00", (Incremental, Fractional), SignedPercentage_Incremental_Fractional),
+                    ]
                 },
-
-                #[doc = "Servo position feedback sensor for monitoring actuator positions. Absolute (instant) change. Neurons encoded fractionally exponentially"]
-                ServoPositionAbsoluteFractional => {
-                    friendly_name: "Servo Position (Absolute Change, Fractional Encoding)",
-                    snake_case_identifier: "servo_position_absolute_fractional",
-                    base_ascii: b"iSVP00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..33),
-                    default_coder_type: SignedPercentage_Absolute_Fractional,
-                    wrapped_data_type: WrappedIOType::SignedPercentage,
-                    data_type: SignedPercentage,
-                },
-
-                #[doc = "Servo position feedback sensor for monitoring actuator positions. Incremental change. Neurons encoded linearly"]
-                ServoPositionIncrementalLinear => {
-                    friendly_name: "Servo Position (Incremental Change, Linear Encoding)",
-                    snake_case_identifier: "servo_position_incremental_linear",
-                    base_ascii: b"Isvp00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..u32::MAX),
-                    default_coder_type: SignedPercentage_Incremental_Linear,
-                    wrapped_data_type: WrappedIOType::SignedPercentage,
-                    data_type: SignedPercentage,
-                },
-
-                #[doc = "Servo position feedback sensor for monitoring actuator positions. Incremental change. Neurons encoded fractionally exponentially"]
-                ServoPositionIncrementalFractional => {
-                    friendly_name: "Servo Position (Incremental Change, Fractional Encoding)",
-                    snake_case_identifier: "servo_position_incremental_fractional",
-                    base_ascii: b"ISVP00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..33),
-                    default_coder_type: SignedPercentage_Incremental_Fractional,
-                    wrapped_data_type: WrappedIOType::SignedPercentage,
-                    data_type: SignedPercentage,
-                },
-
-
 
 
                 #[doc = "Servo motion feedback sensor for monitoring actuator positions. Absolute (instant) change. Neurons encoded linearly"]
@@ -400,39 +171,12 @@ macro_rules! sensor_definition {
                     default_coder_type: SignedPercentage_Absolute_Linear,
                     wrapped_data_type: WrappedIOType::SignedPercentage,
                     data_type: SignedPercentage,
-                },
-
-                #[doc = "Servo motion feedback sensor for monitoring actuator positions. Absolute (instant) change. Neurons encoded fractionally exponentially"]
-                ServoMotionAbsoluteFractional => {
-                    friendly_name: "Servo Motion (Absolute Change, Fractional Encoding)",
-                    snake_case_identifier: "servo_motion_absolute_fractional",
-                    base_ascii: b"iSVM00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..33),
-                    default_coder_type: SignedPercentage_Absolute_Fractional,
-                    wrapped_data_type: WrappedIOType::SignedPercentage,
-                    data_type: SignedPercentage,
-                },
-
-                #[doc = "Servo motion feedback sensor for monitoring actuator positions. Incremental change. Neurons encoded linearly"]
-                ServoMotionIncrementalLinear => {
-                    friendly_name: "Servo Motion (Incremental Change, Linear Encoding)",
-                    snake_case_identifier: "servo_motion_incremental_linear",
-                    base_ascii: b"Isvm00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..u32::MAX),
-                    default_coder_type: SignedPercentage_Incremental_Linear,
-                    wrapped_data_type: WrappedIOType::SignedPercentage,
-                    data_type: SignedPercentage,
-                },
-
-                #[doc = "Servo motion feedback sensor for monitoring actuator positions. Incremental change. Neurons encoded fractionally exponentially"]
-                ServoMotionIncrementalFractional => {
-                    friendly_name: "Servo Motion (Incremental Change, Fractional Encoding)",
-                    snake_case_identifier: "servo_motion_incremental_fractional",
-                    base_ascii: b"ISVM00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..2, 1..2, 1..33),
-                    default_coder_type: SignedPercentage_Incremental_Fractional,
-                    wrapped_data_type: WrappedIOType::SignedPercentage,
-                    data_type: SignedPercentage,
+                    subtype_mappings: [
+                        (b"isvm00", (Absolute, Linear), SignedPercentage_Absolute_Linear),
+                        (b"iSVM00", (Absolute, Fractional), SignedPercentage_Absolute_Fractional),
+                        (b"Isvm00", (Incremental, Linear), SignedPercentage_Incremental_Linear),
+                        (b"ISVM00", (Incremental, Fractional), SignedPercentage_Incremental_Fractional),
+                    ]
                 },
 
                 //endregion
@@ -448,17 +192,10 @@ macro_rules! sensor_definition {
                     default_coder_type: MiscData_Absolute,
                     wrapped_data_type: WrappedIOType::MiscData(None),
                     data_type: MiscData,
-                },
-
-                #[doc = "Miscellaneous area for all types of data FEAGI has no specific implementation for. Incremental change."]
-                MiscellaneousIncremental => {
-                    friendly_name: "Miscellaneous (incremental_change)",
-                    snake_case_identifier: "miscellaneous_incremental",
-                    base_ascii: b"imis00",
-                    channel_dimension_range: CorticalChannelDimensionRange::new(1..u32::MAX, 1..u32::MAX, 1..u32::MAX),
-                    default_coder_type: MiscData_Incremental,
-                    wrapped_data_type: WrappedIOType::MiscData(None),
-                    data_type: MiscData,
+                    subtype_mappings: [
+                        (b"imis00", (Absolute), MiscData_Absolute),
+                        (b"Imis00", (Incremental), MiscData_Incremental),
+                    ]
                 },
 
                 //endregion
